+++ optashift-employee-rostering-gwtui/src/test/java/org/optaplanner/openshift/employeerostering/gwtui/client/pages/rotation/Infinite60MinutesScaleTest.java
@@ -24,20 +24,20 @@ public class Infinite60MinutesScaleTest {
         final Infinite60MinutesScale scale = new Infinite60MinutesScale(start, end);
 
         // Boundaries
-        Assert.assertEquals((Double) startInGridPixels, scale.toGridPixels(start), ACCEPTED_ERROR);
-        Assert.assertEquals((Double) endInGridPixels, scale.toGridPixels(end), ACCEPTED_ERROR);
+        Assert.assertEquals(startInGridPixels, scale.toGridPixels(start), ACCEPTED_ERROR);
+        Assert.assertEquals(endInGridPixels, scale.toGridPixels(end), ACCEPTED_ERROR);
         Assert.assertEquals(start, scale.toScaleUnits(startInGridPixels));
         Assert.assertEquals(end, scale.toScaleUnits(endInGridPixels));
 
         // Over-boundaries
-        Assert.assertEquals((Double) (-1d), scale.toGridPixels(start.minus(60, MINUTES)), ACCEPTED_ERROR);
-        Assert.assertEquals((Double) (endInGridPixels + 1d), scale.toGridPixels(end.plus(60, MINUTES)), ACCEPTED_ERROR);
-        Assert.assertEquals(start.minusMinutes(60L), scale.toScaleUnits(startInGridPixels - 1L));
-        Assert.assertEquals(end.plusMinutes(60L), scale.toScaleUnits(endInGridPixels + 1));
+        Assert.assertEquals(-1d, scale.toGridPixels(start.minus(60, MINUTES)), ACCEPTED_ERROR);
+        Assert.assertEquals(endInGridPixels + 1d, scale.toGridPixels(end.plus(60, MINUTES)), ACCEPTED_ERROR);
+        Assert.assertEquals(start.minusMinutes(60), scale.toScaleUnits(startInGridPixels - 1));
+        Assert.assertEquals(end.plusMinutes(60), scale.toScaleUnits(endInGridPixels + 1));
 
         // 3 Days and 4 Hours after
         final LocalDateTime d = start.plus(3, DAYS).plus(4, HOURS);
-        Assert.assertEquals((Double) (24d * 3 + 4d), scale.toGridPixels(d), ACCEPTED_ERROR);
+        Assert.assertEquals(24d * 3 + 4d, scale.toGridPixels(d), ACCEPTED_ERROR);
         Assert.assertEquals(d, scale.toScaleUnits(24d * 3 + 4d));
     }
 }
